feat: price bigint migration, auth hardening, i18n, forgot/reset password & admin improvements
Backend:
- Migrate all price/wallet fields from decimal to long (bigint) for VND precision
- Standardize pagination across all listing endpoints with PagedResponse<T>
- Enable JWT issuer/audience validation with zero clock skew
- Implement forgot password & reset password flows
- Add current password verification on profile updates
- Add EF migrations for price conversion and pagination seed data

Frontend:
- Add i18n support with Vietnamese (vi) and English (en) locales
- Add ForgotPassword, ResetPassword, and Profile pages
- Add LanguageSwitcher, PageSkeleton, GameCardSkeleton components
- Overhaul admin panel: sorting, filtering, and improved UX across all tabs
- Update all major pages (cart, payment, store, game detail, home, library, wishlist)
- Enhanced navbar with language toggle and improved navigation

Infrastructure:
- Update .gitignore with broader exclusions
- Rewrite README with complete setup and API documentation
- Enhance run-all.sh and kill-all.sh with argument parsing and logging
- Clean up project dump files

+++ GameStore.DTOs/Admin/AdminUserUpdateDto.cs
@@ -12,6 +12,6 @@ public class AdminUserUpdateDto
     public string? Email { get; set; }
     public string? Phone { get; set; }
     public string? AvatarUrl { get; set; }
-    public decimal? Wallet { get; set; }
+    public long? Wallet { get; set; }
     public bool? IsActive { get; set; }
 }
