+++ GameStore.DTOs/Orders/OrderHistoryDto.cs
@@ -9,7 +9,7 @@ public class OrderHistoryDto
 {
     public int Id { get; set; }
     public DateTime OrderDate { get; set; }
-    public decimal TotalAmount { get; set; }
+    public long TotalAmount { get; set; }
     public string Status { get; set; } = string.Empty;
     public string PaymentMethod { get; set; } = string.Empty;
     public List<OrderHistoryItemDto> Items { get; set; } = new();
@@ -20,5 +20,5 @@ public class OrderHistoryItemDto
     public int GameId { get; set; }
     public string GameTitle { get; set; } = string.Empty;
     public int Quantity { get; set; }
-    public decimal UnitPrice { get; set; }
+    public long UnitPrice { get; set; }
 }
