+++ GameStore.Services/GameService.cs
@@ -23,7 +23,7 @@ public class GameService : IGameService
     public async Task<List<Game>> GetByGenre(int genreId) => await _gameRepository.GetByGenreAsync(genreId);
 
     public async Task<(List<Game> Games, int TotalCount)> Search(string? keyword, int? genreId,
-        decimal? minPrice, decimal? maxPrice, string? sortBy, bool descending, int page, int pageSize)
+        long? minPrice, long? maxPrice, string? sortBy, bool descending, int page, int pageSize)
     {
         return await _gameRepository.SearchAsync(keyword, genreId, minPrice, maxPrice, sortBy, descending, page, pageSize);
     }
